% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ArArRedux.R
\name{average}
\alias{average}
\title{Calculate the arithmetic mean}
\usage{
average(x, i = NULL, newlabel = NULL)
}
\arguments{
\item{x}{an object of class \code{redux} or \code{logratios}}

\item{i}{(optional) vector of sample indices}

\item{newlabel}{(optional) string with the new label to be assigned
to the averaged values}
}
\value{
an object of the same class as \code{x}
}
\description{
Calculate the arithmetic mean of some logratio data
}
\examples{
data(Melbourne)
K <- average(Melbourne$X,grep("K:",Melbourne$X$labels),newlabel="K-glass")
plotcorr(K)
}

