% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\docType{class}
\name{blankcorrected}
\alias{blankcorrected}
\title{The \code{blankcorrected} class}
\description{
An object class containing blank-corrected mass spectrometry data
}
\details{
Extends the class classes \code{\link{timeresolved}} and
\code{\link{PHdata}} by adding an additional list item
\code{blankindices} containg the index of the nearest
blank. \code{\link{fitlogratios}} uses this information to group
the samples during regression to 'time zero'.
}

