% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\alias{calibration.default}
\alias{calibration.WiscAr}
\title{Detector calibration}
\usage{
calibration(x, ...)

\method{calibration}{default}(x, clabel, ...)

\method{calibration}{WiscAr}(x, clabel, ...)
}
\arguments{
\item{x}{an object of class \code{redux}}

\item{...}{optional arguments}

\item{clabel}{the label of the detector calibration data}
}
\value{
an object of class \code{redux}
}
\description{
Apply the detector calibration for multicollector data
}
\examples{
data(Melbourne)
C <- calibration(Melbourne$X,"DCAL")
plotcorr(C)
}
