% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cl.R
\name{clcorrection}
\alias{clcorrection}
\title{Cl-interference correction}
\usage{
clcorrection(X, irr)
}
\arguments{
\item{X}{an object of class \code{redux}}

\item{irr}{the irradiation schedule}
}
\value{
an object of class \code{redux}
}
\description{
Apply the interference correction for the Cl-decay products
}
\examples{
data(Melbourne)
Cl <- clcorrection(Melbourne$X,Melbourne$irr)
plotcorr(Cl)
}
