% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoratios.R
\name{isoratios}
\alias{isoratios}
\title{Isochron ratios}
\usage{
isoratios(X, irr, fract = NULL, ca = NULL, k = NULL, inverse = TRUE)
}
\arguments{
\item{X}{an object of class \code{redux} containing some
interference corrected logratio intercept data}

\item{irr}{the irradiation schedule}

\item{fract}{list with air shot data (one item per denominator
isotope)}

\item{ca}{an object of class \code{\link{logratios}} with
Ca-interference data (not necessary if interferences are
included in X)}

\item{k}{an object of class \code{\link{logratios}} with
K-interference data (not necessary if interferences are
included in X)}

\item{inverse}{logical. If \code{TRUE}, returns inverse isochron
ratios (\eqn{^{36}}Ar/\eqn{^{40}}Ar
vs. \eqn{^{39}}Ar/\eqn{^{40}}Ar). Otherwise returns normal
isochron ratios (\eqn{^{40}}Ar/\eqn{^{36}}Ar
vs. \eqn{^{39}}Ar/\eqn{^{36}}Ar).}
}
\value{
an object of class \code{link{isoratios}} containing the
    isochron ratios and their covariance matrix.
}
\description{
Calculate the Calculate the inverse isochron ratios of interference
corrected logratio intercept data
}
\examples{
data(Melbourne)
IR <- isoratios(Melbourne$X,irr=Melbourne$irr,fract=Melbourne$fract)
}
