% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{loaddata}
\alias{loaddata}
\title{Load mass spectrometer data}
\usage{
loaddata(fid, MS = "ARGUS-VI")
}
\arguments{
\item{fid}{the file or directory name containing the data}

\item{MS}{one of either \code{ARGUS-VI} or \code{WiscAr}}
}
\value{
an object of class \code{timeresolved} or \code{PHdata} and
    \code{ArgusVI} or \code{WiscAr}
}
\description{
Loads a file or directory with raw mass spectrometer data
}
\examples{
samplefile <- system.file("Samples.csv",package="ArArRedux")
masses <- c("Ar37","Ar38","Ar39","Ar40","Ar36")
m <- loaddata(samplefile,MS='ARGUS-VI') # samples and J-standards
plot(m,"MD2-1a","Ar40")
}
