% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\docType{class}
\name{logratios}
\alias{logratios}
\title{The \code{logratios} class}
\description{
An object class containing logratio intercepts
}
\details{
A list with the following items:

\code{labels}: a vector of strings denoting the names of the runs\cr
\code{num}: a vector of strings denoting the numerator isotopes\cr
\code{den}: a vector of strings denoting the denominator isotopes\cr
\code{intercepts}: a vector of logratio intercepts or values\cr
\code{covmat}: the covariance matrix of \code{intercepts}\cr
\code{irr}: a vector of strings denoting the irradiation runs\cr
\code{pos}: a vector of integers with the positions in the
irradiation stack \cr
\code{thedate}: a vector containing the acquisition dates and times\cr
\code{nlr}: a vector with the number of logratios per run\cr
}

