% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{Melbourne}
\alias{Melbourne}
\title{An example dataset}
\description{
Contains all the relevant information needed for the data reduction
some ARGUS-IV data from the University of Melbourne
}
\examples{
data(Melbourne)
plotcorr(Melbourne$X)
}
\author{
David Philips \email{dphillip@unimelb.edu.au}
}
