% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ArArRedux.R
\name{newredux}
\alias{newredux}
\title{Create a new \code{\link{redux}} object}
\usage{
newredux(X, Jpos, detectors = list(Ar36 = "H1", Ar37 = "L2", Ar38 = "L1", Ar39
  = "AX", Ar40 = "H1"))
}
\arguments{
\item{X}{an object of class \code{\link{logratios}}}

\item{Jpos}{a vector of integers denoting the positions of the
fluence monitors in the irradiation stack}

\item{detectors}{a list of strings denoting the detectors for each
argon isotope}
}
\value{
an object of class \code{\link{redux}}
}
\description{
Initialises a new \code{\link{redux}} object by packing a
\code{\link{logratios}} dataset together with all the parameters
needed for age calculation
}

