% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{class}
\name{PHdata}
\alias{PHdata}
\title{The \code{PHdata} class}
\description{
An object class containing time resolved 'peak-hopping' mass
spectrometry data
}
\details{
A list with the following items:

\code{masses}: a vector of strings denoting the isotopes monitored in each run\cr
\code{signals}: a list with objects of class \code{\link{timeresolved}}, each corresponding
to a detector (i.e. length(\code{signals})==1 for single collector instruments).
}
\seealso{
\code{\link{loaddata}}
}

