% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.timeresolved}
\alias{plot.timeresolved}
\alias{plot.WiscAr}
\alias{plot.PHdata}
\title{Plot a time resolved mass spectrometry signal}
\usage{
\method{plot}{timeresolved}(x, label = NULL, mass = NA, run = 1, hop = "101", tmin = 0, ...)

\method{plot}{WiscAr}(x, hop = "101", ...)

\method{plot}{PHdata}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{timeresolved}} or
\code{\link{PHdata}}}

\item{label}{a string with the name of the run}

\item{mass}{a string indicating the isotope of interest}

\item{run}{the run number}

\item{hop}{identifier for mass channel in Noblesse instruments (if
\code{x} has class \code{WiscAr}).}

\item{tmin}{time stamp (in seconds) of 'time zero'}

\item{...}{optional parameters}
}
\description{
Plots the raw signal of a given isotope against time.
}
\examples{
samplefile <- system.file("Samples.csv",package="ArArRedux")
mMC <- loaddata(samplefile)
plot(mMC,"MD2-1a")
mPH <- loaddata(samplefile)
plot(mPH,"MD2-1a","Ar40")
}
