% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotcorr}
\alias{plotcorr}
\title{Plot a matrix with correlation coefficients}
\usage{
plotcorr(X)
}
\arguments{
\item{X}{a data structure (list) containing an item called `covmat' (covariance matrix)}
}
\description{
Converts the covariance matrix to a correlation matrix and plots
this is a coloured image for visual inspection.
}
\examples{
data(Melbourne)
plotcorr(Melbourne$X)
}

