% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read}
\alias{read}
\title{Read mass spectrometer data}
\usage{
read(x, blabel, Jpos, kdat = NULL, cadat = NULL, ddat = NULL, MS = "ARGUS-VI")
}
\arguments{
\item{x}{a \code{.csv} file with samples and fluence monitor data
(if \code{MS = 'ARGUS-VI'}) OR a directory with \code{.RUN}
files (if \code{MS = 'WiscAr'})}

\item{blabel}{a prefix string denoting the blanks}

\item{Jpos}{a vector of integers denoting the positions of the
fluence monitors in the irradiation stack}

\item{kdat}{a \code{.csv} file with the K-interference monitor data
(if \code{MS = 'ARGUS-VI'}) OR a directory with \code{.RUN}
files of K-interference monitor data (if \code{MS = 'WiscAr'})
(optional)}

\item{cadat}{a \code{.csv} file with the Ca-interference monitor
data (if \code{MS = 'ARGUS-VI'}) OR a directory with
\code{.RUN} files of Ca-interference monitor data (if \code{MS
= 'WiscAr'}) (optional)}

\item{ddat}{a \code{.csv} file with the detector calibration data
(if \code{MS = 'ARGUS-VI'}) OR the prefix of the standard gas
measurements (if \code{MS = 'WiscAr'}) (optional)}

\item{MS}{a string denoting the type of mass spectrometer. At the
moment there are two options: \code{'ARGUS-VI'}, for data that
are acquired on the eponymous instrument following the
measurement protocols at Melbourne University, and
\code{'WiscAr'} for the 5-detector Noblesse instrument at the
University of Wisconsin. Please contact the author to add other
file formats to Ar-Ar_Redux.}
}
\value{
an object of class \code{\link{redux}}.
}
\description{
Reads raw mass spectrometer data and parses it into a
\code{\link{redux}} format for further processing.
}
\examples{
samplefile <- system.file("Samples.csv",package="ArArRedux")
kfile <- system.file("K-glass.csv",package="ArArRedux")
cafile <- system.file("Ca-salt.csv",package="ArArRedux")
dfile <- system.file("Calibration.csv",package="ArArRedux")
X <- read(samplefile,blabel="EXB#",Jpos=c(3,15),
          kdat=kfile,cadat=cafile,ddat=dfile,MS='ARGUS-VI')
plotcorr(X)
}
