% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{class}
\name{redux}
\alias{redux}
\title{The \code{redux} class}
\description{
An object class that is used throughout Ar-Ar_Redux
}
\details{
A list with the following items:

\code{labels}: a vector of strings denoting the names of the runs\cr
\code{num}: a vector of strings denoting the numerator isotopes\cr
\code{den}: a vector of strings denoting the denominator isotopes\cr
\code{intercepts}: a vector of logratio intercepts or values\cr
\code{covmat}: the covariance matrix of \code{intercepts}\cr
\code{irr}: a vector of strings denoting the irradiation runs\cr
\code{pos}: a vector of integers with the positions in the
irradiation stack \cr
\code{thedate}: a vector containing the acquisition dates and times\cr
\code{nlr}: a vector with the number of logratios per run\cr
\code{param}: a list of global parameters
}
\seealso{
param
}
