% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/documentation.R
\docType{class}
\name{results}
\alias{results}
\title{The \code{results} class}
\description{
A list with the following items:
}
\details{
\code{labels}: a vector of strings denoting the names of the runs\cr
\code{intercepts}: a vector of ages\cr
\code{covmat}: the covariance matrix of \code{intercepts}\cr
\code{thedate}: a vector containing the acquisition dates and times\cr
}

