% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{summary.results}
\alias{summary.results}
\title{Summary table}
\usage{
\method{summary}{results}(object, ...)
}
\arguments{
\item{object}{an objct of class \code{\link{results}}}

\item{...}{no other arguments}
}
\description{
Plots the ages and their standard errors
}
\examples{
data(Melbourne)
ages <- process(Melbourne$X,Melbourne$irr,Melbourne$fract)
summary(ages)[1:5,]
}

