% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{class}
\name{timeresolved}
\alias{timeresolved}
\title{The \code{timeresolved} class}
\description{
An object class containing time resolved multi-collector mass spectrometry data
}
\details{
A list with the following items:

\code{masses}: a vector of strings denoting the isotopes monitored in each run\cr
\code{irr}: a vector of strings denoting the irradiation runs\cr
\code{pos}: a vector of integers with the positions in the
irradiation stack \cr
\code{thedate}: a vector containing the acquisition dates and times\cr
\code{d}: a data table\cr
\code{thetime}: a matrix with the measurement times
}
\seealso{
\code{\link{loaddata}}
}
