% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ArArRedux.R
\name{calibration}
\alias{calibration}
\title{Detector calibration}
\usage{
calibration(X, clabel)
}
\arguments{
\item{X}{an object of class \code{redux}}

\item{clabel}{the label of the detector calibration data}
}
\value{
an object of class \code{redux}
}
\description{
Apply the detector calibration for multicollector data
}
\examples{
data(Melbourne)
C <- calibration(Melbourne$X,"DCAL")
plotcorr(C)
}

